refactor: Renamed CanvasRecorder
