+++ src/colorizer/recorders/CanvasRecorder.ts
